requirements/requirements.txt
Django==4.2.2
psycopg2-binary
django-environ
djangorestframework
django-oauth-toolkit
